@@ -1,20 +1,24 @@
 {
   "name": "@repo/eslint-config",
   "version": "0.0.0",
+  "type": "module",
   "private": true,
-  "files": [
-    "library.js",
-    "next.js",
-    "react-internal.js",
-    "server.js"
-  ],
+  "exports": {
+    "./base": "./base.js",
+    "./next-js": "./next.js",
+    "./react-internal": "./react-internal.js"
+  },
   "devDependencies": {
-    "@vercel/style-guide": "^6.0.0",
-    "eslint-config-turbo": "^2.3.0",
-    "eslint-config-prettier": "^9.1.0",
+    "@eslint/js": "^9.39.1",
+    "@next/eslint-plugin-next": "^15.5.0",
+    "eslint": "^9.39.1",
+    "eslint-config-prettier": "^10.1.1",
     "eslint-plugin-only-warn": "^1.1.0",
-    "@typescript-eslint/parser": "^8.15.0",
-    "@typescript-eslint/eslint-plugin": "^8.15.0",
-    "typescript": "^5.8.3"
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-turbo": "^2.7.1",
+    "globals": "^16.5.0",
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.50.0"
   }
 }
